Add app1 marker and tests
